+++ Client/src/app/components/transmilenio/transmilenio-list/transmilenio-list.component.ts
@@ -3,6 +3,7 @@ import { Conductor } from 'src/app/model/conductor/conductor';
 import { Transmilenio } from 'src/app/model/transmilenio/transmilenio';
 import { ConductorService } from 'src/app/services/conductor/conductor.service';
 import { TransmilenioService } from 'src/app/services/transmilenio/transmilenio.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-transmilenio-list',
@@ -29,26 +30,48 @@ export class TransmilenioListComponent implements OnInit {
   }
 
   public eliminarTransmilenio(id: number) {
-    let transmilenio: Transmilenio = new Transmilenio();
-    let conductoresTransmilenio: Conductor[] = [];
-    this.transmilenioService.findAll().subscribe(rTransmilenios => {
-      this.transmilenios = rTransmilenios;
-      this.transmilenioService.findById(id).subscribe(rTransmilenio => {
-        transmilenio = rTransmilenio;
-        for (let conductor of this.conductores) {
-          if (this.contains(conductor, transmilenio)) {
-            conductoresTransmilenio.push(conductor);
-          }
-        }
-        if (conductoresTransmilenio.length >= 1) {
-          window.alert(`El transmilenio ${transmilenio.placa} no se puede eliminar porque tiene conductores asignados.`);
-        } else {
-          this.transmilenioService.delete(id).subscribe(response => {
-            this.transmilenios.splice(this.getIndex(this.transmilenios, transmilenio), 1);
+    Swal.fire({
+      title: '¿Seguro de Eliminarlo?',
+      text: "Eliminar este Bus",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Sí'
+    }).then((result) => {
+        let transmilenio: Transmilenio = new Transmilenio();
+        let conductoresTransmilenio: Conductor[] = [];
+        this.transmilenioService.findAll().subscribe(rTransmilenios => {
+          this.transmilenios = rTransmilenios;
+          this.transmilenioService.findById(id).subscribe(rTransmilenio => {
+            transmilenio = rTransmilenio;
+            for (let conductor of this.conductores) {
+              if (this.contains(conductor, transmilenio)) {
+                conductoresTransmilenio.push(conductor);
+              }
+            }
+            if (conductoresTransmilenio.length >= 1) {
+              window.alert(`El transmilenio ${transmilenio.placa} no se puede eliminar porque tiene conductores asignados.`);
+              Swal.fire({
+                icon: 'error',
+                title: 'No Eliminado',
+                text: 'El bus' + transmilenio.placa +' no se puede eliminar porque tiene conductores asignados.'
+              })  
+            } else {
+                this.transmilenioService.delete(id).subscribe(response => {
+                this.transmilenios.splice(this.getIndex(this.transmilenios, transmilenio), 1);
+                Swal.fire({
+                  position: 'top-end',
+                  icon: 'success',
+                  title: 'Bus Eliminado',
+                  showConfirmButton: false,
+                  timer: 1500
+                })
+              });
+            }
           });
-        }
-      });
-    });
+        });
+    })
   }
 
   public contains(conductor: Conductor, transmilenio: Transmilenio): boolean {
